requirements.txt
appdirs==1.4.4
atomicwrites==1.4.1
attrs==22.1.0
colorama==0.4.5
coverage==6.4.3
distlib==0.3.5
filelock==3.7.1
get-nic==0.0.3
iniconfig==1.1.1
lxml==4.9.1
nose2==0.12.0
packaging==21.3
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
pyparsing==3.0.9
pyshark==0.5.3
pytest==7.1.2
python-dotenv==0.20.0
six==1.16.0
toml==0.10.2
tomli==2.0.1
tox==3.25.1
virtualenv==20.16.3
